Increase ibrowse max sessions and disable request pipelining
Set the global ibrowse max sessions for a single host to 256 (default
value is 10). This prevents mini_s3 requests to bookshelf from
erroring out when many requests are inflight.

Also disable request pipelining for ibrowse requests. Unclear if
pipelining would give us much benefit to begin with and some of the
error traces we observed suggested that pipelining was involved --
where a request was sent on a connection that the server was closing.