Automated script for generating tech-preview wiki page
This automates the generation of timed-expiration S3 links to various
tech preview packages for all our products.  Given AWS credentials, it
will generate the complete Confluence wiki markup to replace the
entire contents of
http://wiki.corp.opscode.com/display/CORP/Private+Chef+11+Preview+Builds
with new links that are good for one week.

It is designed to be simple to maintain and use, since doing this
stuff by hand is a special kind of hell.

Although this deals with packages other than just Private Chef,
putting this code in this repository seemed better than creating a new
repository for it (too heavy-weight) or sticking it in a Gist (easy to
"lose").