Add PostgreSQL 9.1 and 9.2 definitions
These will supercede the old 'postgresql' definition from
opscode-software (which is "unversioned").  OPC will be using 9.2 from
this point onward, but we also need to keep the 9.1 binaries around in
order to be able to upgrade existing installations (see the
documentation for PostgreSQL's 'pg_upgrade' tool for more).

To facilitate this, we begin here to separate installed PostgreSQL
artifacts into directories based on the PostgreSQL release version.
This will allow multiple versions to exist alongside each other in a
manageable way.