Actually ensure that old cookbooks get migrated on HA systems
The attribute by which we find old cookbooks to move them into
bookshelf is under the "opscode_chef" key, which we need to explicitly
mark as being added to the hash of attributes we subsequently load
into the node.

Since we programmatically manipulate this path on HA systems, it is
important that this new value gets written into the node; otherwise,
the value from attributes/default.rb is used, which is incorrect for
HA systems.