[don't merge] darklaunch - set couchdb_{containers,groups} true
This commit is obviated by work on the ec11-2-0/int branch.
It is not intended to be merged and is only to be used for
testing the solr4 work.

The reason we need to set these attributes is because org-creator
doesn't go through the load balancer to talk to opscode-account
for org creation. If there are no xdarklaunch headers present,
opscode-account will read the dark_launch.json file off disk
that is rendered from the private_chef::default recipe. If a
darklaunch value is not present in said file, then account
assumes that value is false, hence couchdb_containers would
be false and account would try to create containers in SQL,
breaking the desired behavior.