Configure paths for applications with customha topology
When creating the customha topology, I stripped out everything to do with drbd
(at least when you choose the customha topology, the ha topology remains
unchanged). Unfortunately I stripped out a little too much, including the
parts that configure all data to be stored in the HA block device. This commit
fixes that, configuring the paths for the customha topology also.

Currently these paths are hardcoded, and contain the word 'drbd'. At some
point in future this may change for the customha topology to not have 'drbd'
in the name and/or be customizable by the end user.