Default to `lucid` for debian family
We build packages on Ubuntu 10.04 (lucid), then test them on later
versions. However, not all new versions are tested; currently only
Ubuntu, and up to version 12.04. By defaulting to `lucid` we allow
users to install on any debian-family platform, and we print a log
message (or have a resource show in the output formatter).

Long term we should fix the issue (#460), where we have tester systems
for other Ubuntu versions - in particular Ubuntu 14.04 as it's an
LTS.